Cassowary: deprecate the use of `|` for strength
Instead of using the `|` on the constraint to add the strength on the
constraint, change the `Solver` to make the strength parameter a labelled
parameter.  There were previously three different ways to add a constraint with
a strength:

1. `solver.add(constraint:_:)`
2. `solver.add(constraint: Constraint(_:strength:))` or `solver.add(constraint: Constraint(_:_:_:))`
3. `solver.add(constraint:)`, `|` on the constraint to compose the strength

This effectively removes `solver.add(constraint:_:)` in favour of
`solver.add(constraint:strength:)` and retains option 2.  This way, the strength
can be specified when adding the constraint to the solver or when creating the
constraint.